GISProject/requirements.txt
h5py~=3.6.0
pyhdf~=0.10.3
numpy~=1.22.3
pillow~=9.0.1